ConDec-446: Move Comment and Sentence model interfaces and classes to upper model package
