+++ doc/links.ucls
@@ -109,7 +109,7 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <interface id="11" language="java" name="de.uhd.ifi.se.decision.management.jira.extraction.model.Comment" 
+  <interface id="11" language="java" name="de.uhd.ifi.se.decision.management.jira.model.Comment" 
     project="git" file="/git/src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/model/Comment.java" 
     binary="false" corner="BOTTOM_RIGHT">    
     <position height="-1" width="-1" x="600" y="-494"/>    
@@ -119,7 +119,7 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </interface>  
-  <interface id="12" language="java" name="de.uhd.ifi.se.decision.management.jira.extraction.model.Sentence" 
+  <interface id="12" language="java" name="de.uhd.ifi.se.decision.management.jira.model.Sentence" 
     project="git" file="/git/src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/model/Sentence.java" 
     binary="false" corner="BOTTOM_RIGHT">    
     <position height="-1" width="-1" x="1276" y="-374"/>    
@@ -129,7 +129,7 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </interface>  
-  <class id="13" language="java" name="de.uhd.ifi.se.decision.management.jira.extraction.model.impl.CommentImpl" 
+  <class id="13" language="java" name="de.uhd.ifi.se.decision.management.jira.model.impl.CommentImpl" 
     project="git" 
     file="/git/src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/model/impl/CommentImpl.java" 
     binary="false" corner="BOTTOM_RIGHT">    
@@ -140,7 +140,7 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <class id="14" language="java" name="de.uhd.ifi.se.decision.management.jira.extraction.model.impl.SentenceImpl" 
+  <class id="14" language="java" name="de.uhd.ifi.se.decision.management.jira.model.impl.SentenceImpl" 
     project="git" 
     file="/git/src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/model/impl/SentenceImpl.java" 
     binary="false" corner="BOTTOM_RIGHT">
