+++ src/main/java/de/uhd/ifi/se/decision/management/jira/model/KnowledgeType.java
@@ -11,7 +11,6 @@ import com.atlassian.jira.issue.Issue;
 import com.atlassian.jira.issue.IssueManager;
 
 import de.uhd.ifi.se.decision.management.jira.ComponentGetter;
-import de.uhd.ifi.se.decision.management.jira.extraction.model.Sentence;
 
 /**
  * Type of decision knowledge element
