+++ src/main/java/de/uhd/ifi/se/decision/management/jira/view/GraphFiltering.java
@@ -3,8 +3,9 @@ package de.uhd.ifi.se.decision.management.jira.view;
 import com.atlassian.jira.jql.builder.JqlClauseBuilder;
 import com.atlassian.jira.jql.builder.JqlQueryBuilder;
 import com.atlassian.query.clause.Clause;
-import de.uhd.ifi.se.decision.management.jira.extraction.model.Sentence;
+
 import de.uhd.ifi.se.decision.management.jira.model.DecisionKnowledgeElement;
+import de.uhd.ifi.se.decision.management.jira.model.Sentence;
 import de.uhd.ifi.se.decision.management.jira.model.impl.DecisionKnowledgeElementImpl;
 import de.uhd.ifi.se.decision.management.jira.persistence.JiraIssueCommentPersistenceManager;
